Move Dumper logo to utils
