+++ utils.py
@@ -1,5 +1,6 @@
 from __future__ import print_function
 
+import struct
 from os.path import exists
 import sys, os, errno, tempfile
 import common, blockimgdiff, sparse_img
@@ -94,3 +95,102 @@ class vbpatch:
 
     def disavb(self):
         self.patchvb(b'\x02')
+
+
+class DUMPCFG:
+    blksz = 0x1 << 0xc
+    headoff = 0x4000
+    magic = b"LOGO!!!!"
+    imgnum = 0
+    imgblkoffs = []
+    imgblkszs = []
+
+
+class BMPHEAD(object):
+    def __init__(self, buf: bytes = None):  # Read bytes buf and use this struct to parse
+        assert buf is not None, f"buf Should be bytes not {type(buf)}"
+        # print(buf)
+        self.structstr = "<H6I"
+        (
+            self.magic,
+            self.fsize,
+            self.reserved,
+            self.hsize,
+            self.dib,
+            self.width,
+            self.height,
+        ) = struct.unpack(self.structstr, buf)
+
+
+class XIAOMI_BLKSTRUCT(object):
+    def __init__(self, buf: bytes):
+        self.structstr = "2I"
+        (
+            self.imgoff,
+            self.blksz,
+        ) = struct.unpack(self.structstr, buf)
+
+
+class LOGODUMPER(object):
+    def __init__(self, img: str, out: str, dir: str = "pic"):
+        self.out = out
+        self.img = img
+        self.dir = dir
+        self.structstr = "<8s"
+        self.cfg = DUMPCFG()
+        self.chkimg(img)
+
+    def chkimg(self, img: str):
+        assert os.access(img, os.F_OK), f"{img} does not found!"
+        with open(img, 'rb') as f:
+            f.seek(self.cfg.headoff, 0)
+            self.magic = struct.unpack(
+                self.structstr, f.read(struct.calcsize(self.structstr))
+            )[0]
+            while True:
+                m = XIAOMI_BLKSTRUCT(f.read(8))
+                if m.imgoff != 0:
+                    # print(blksz<<0xc)
+                    self.cfg.imgblkszs.append(m.blksz << 0xc)
+                    self.cfg.imgblkoffs.append(m.imgoff << 0xc)
+                    self.cfg.imgnum += 1
+                else:
+                    break
+        # print(self.magic)
+        assert self.magic == b"LOGO!!!!", "File does not match xiaomi logo magic!"
+        print("Xiaomi LOGO!!!! format check pass!")
+
+    def unpack(self):
+        with open(self.img, 'rb') as f:
+            print("Unpack:\n"
+                  "BMP\tSize\tWidth\tHeight")
+            for i in range(self.cfg.imgnum):
+                f.seek(self.cfg.imgblkoffs[i], 0)
+                bmph = BMPHEAD(f.read(26))
+                f.seek(self.cfg.imgblkoffs[i], 0)
+                print("%d\t%d\t%d\t%d" % (i, bmph.fsize, bmph.width, bmph.height))
+                with open(os.path.join(self.out, "%d.bmp" % i), 'wb') as o:
+                    o.write(f.read(bmph.fsize))
+            print("\tDone!")
+
+    def repack(self):
+        with open(self.out, 'wb') as o:
+            off = 0x5
+            for i in range(self.cfg.imgnum):
+                print("Write BMP [%d.bmp] at offset 0x%X" % (i, off << 0xc))
+                with open(os.path.join(self.dir, "%d.bmp" % i), 'rb') as b:
+                    bhead = BMPHEAD(b.read(26))
+                    b.seek(0, 0)
+                    self.cfg.imgblkszs[i] = (bhead.fsize >> 0xc) + 1
+                    self.cfg.imgblkoffs[i] = off
+
+                    o.seek(off << 0xc)
+                    o.write(b.read(bhead.fsize))
+
+                    off += self.cfg.imgblkszs[i]
+            o.seek(self.cfg.headoff)
+            o.write(self.magic)
+            for i in range(self.cfg.imgnum):
+                o.write(struct.pack("<I", self.cfg.imgblkoffs[i]))
+                o.write(struct.pack("<I", self.cfg.imgblkszs[i]))
+            print("\tDone!")
